+++ Assets/AppsFlyer/Plugins/iOS/AppsFlyer+AppController.m
@@ -7,10 +7,11 @@
 
 #import <objc/runtime.h>
 #import "UnityAppController.h"
-#if __has_include(<AppsFlyerLib/AppsFlyerTracker.h>)
-#import <AppsFlyerLib/AppsFlyerTracker.h>
+#import "AppsFlyeriOSWrapper.h"
+#if __has_include(<AppsFlyerLib/AppsFlyerLib.h>)
+#import <AppsFlyerLib/AppsFlyerLib.h>
 #else
-#import "AppsFlyerTracker.h"
+#import "AppsFlyerLib.h"
 #endif
 
 
@@ -27,35 +28,38 @@ static IMP __original_openUrl_Imp __unused;
 + (void)load {
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
-        
-        /** <remove comment if you are swizzling applicationDidBecomeActive>
-         Method method1 = class_getInstanceMethod([self class], @selector(applicationDidBecomeActive:));
-         __original_applicationDidBecomeActive_Imp = method_setImplementation(method1, (IMP)__swizzled_applicationDidBecomeActive);
-         */
-        
-        /** <remove comment if you are swizzling applicationDidEnterBackground>
-         Method method2 = class_getInstanceMethod([self class], @selector(applicationDidEnterBackground:));
-         __original_applicationDidEnterBackground_Imp = method_setImplementation(method2, (IMP)__swizzled_applicationDidEnterBackground);
-         */
-        
-        /** <remove comment if you are swizzling didReceiveRemoteNotification>
-         Method method3 = class_getInstanceMethod([self class], @selector(didReceiveRemoteNotification:));
-         __original_didReceiveRemoteNotification_Imp = method_setImplementation(method3, (IMP)__swizzled_didReceiveRemoteNotification);
-         */
 
-        /** <remove comment if you are swizzling openURL>
-         Method method4 = class_getInstanceMethod([self class], @selector(application:openURL:options:));
-         __original_openUrl_Imp = method_setImplementation(method4, (IMP)__swizzled_openURL);
-         */
-
-        /** <remove comment if you are swizzling continueUserActivity>
-        [self swizzleContinueUserActivity:[self class]];
-        */
+        id swizzleFlag = [[NSBundle mainBundle] objectForInfoDictionaryKey:@"AppsFlyerShouldSwizzle"];
+        BOOL shouldSwizzle = swizzleFlag ? [swizzleFlag boolValue] : NO;
         
+        if(shouldSwizzle){
+            
+            Method method1 = class_getInstanceMethod([self class], @selector(applicationDidBecomeActive:));
+            __original_applicationDidBecomeActive_Imp = method_setImplementation(method1, (IMP)__swizzled_applicationDidBecomeActive);
+                    
+            Method method2 = class_getInstanceMethod([self class], @selector(applicationDidEnterBackground:));
+            __original_applicationDidEnterBackground_Imp = method_setImplementation(method2, (IMP)__swizzled_applicationDidEnterBackground);
+            
+           
+            Method method3 = class_getInstanceMethod([self class], @selector(didReceiveRemoteNotification:));
+            __original_didReceiveRemoteNotification_Imp = method_setImplementation(method3, (IMP)__swizzled_didReceiveRemoteNotification);
+            
+           
+            Method method4 = class_getInstanceMethod([self class], @selector(application:openURL:options:));
+            __original_openUrl_Imp = method_setImplementation(method4, (IMP)__swizzled_openURL);
+            
+            if (_AppsFlyerdelegate == nil) {
+                _AppsFlyerdelegate = [[AppsFlyeriOSWarpper alloc] init];
+            }
+
+            [[AppsFlyerLib shared] setDelegate:_AppsFlyerdelegate];
+           
+            [self swizzleContinueUserActivity:[self class]];
+        }
+
     });
 }
 
-/** <remove comment if you are swizzling continueUserActivity>
 +(void)swizzleContinueUserActivity:(Class)class {
     
     SEL originalSelector = @selector(application:continueUserActivity:restorationHandler:);
@@ -71,35 +75,33 @@ static IMP __original_openUrl_Imp __unused;
         class_replaceMethod(class, originalSelector, (IMP)__swizzled_continueUserActivity, method_getTypeEncoding(swizzledMethod));
     }
 }
- 
- BOOL __swizzled_continueUserActivity(id self, SEL _cmd, UIApplication* application, NSUserActivity* userActivity, void (^restorationHandler)(NSArray*)) {
- NSLog(@"swizzled continueUserActivity");
- [[AppsFlyerTracker sharedTracker] continueUserActivity:userActivity restorationHandler:restorationHandler];
- 
- if(__original_continueUserActivity_Imp){
- return ((BOOL(*)(id, SEL, UIApplication*, NSUserActivity*))__original_continueUserActivity_Imp)(self, _cmd, application, userActivity);
- }
- 
- return YES;
- }
- 
-*/
 
-/** <remove comment if you are swizzling applicationDidBecomeActive>
+BOOL __swizzled_continueUserActivity(id self, SEL _cmd, UIApplication* application, NSUserActivity* userActivity, void (^restorationHandler)(NSArray*)) {
+    NSLog(@"swizzled continueUserActivity");
+    [[AppsFlyerLib shared] continueUserActivity:userActivity restorationHandler:restorationHandler];
+    
+    if(__original_continueUserActivity_Imp){
+        return ((BOOL(*)(id, SEL, UIApplication*, NSUserActivity*))__original_continueUserActivity_Imp)(self, _cmd, application, userActivity);
+    }
+    
+    return YES;
+}
+
+
+
 void __swizzled_applicationDidBecomeActive(id self, SEL _cmd, UIApplication* launchOptions) {
     NSLog(@"swizzled applicationDidBecomeActive");
     
-    if(didEnteredBackGround){
-        [[AppsFlyerTracker sharedTracker] trackAppLaunch];
+    if(didEnteredBackGround && AppsFlyeriOSWarpper.didCallStart == YES){
+        [[AppsFlyerLib shared] start];
     }
     
     if(__original_applicationDidBecomeActive_Imp){
         ((void(*)(id,SEL, UIApplication*))__original_applicationDidBecomeActive_Imp)(self, _cmd, launchOptions);
     }
 }
- */
 
-/** <remove comment if you are swizzling applicationDidEnterBackground>
+
 void __swizzled_applicationDidEnterBackground(id self, SEL _cmd, UIApplication* application) {
     NSLog(@"swizzled applicationDidEnterBackground");
     didEnteredBackGround = YES;
@@ -107,34 +109,32 @@ void __swizzled_applicationDidEnterBackground(id self, SEL _cmd, UIApplication* 
         ((void(*)(id,SEL, UIApplication*))__original_applicationDidEnterBackground_Imp)(self, _cmd, application);
     }
 }
- */
 
-/** <remove comment if you are swizzling didReceiveRemoteNotification>
+
 BOOL __swizzled_didReceiveRemoteNotification(id self, SEL _cmd, UIApplication* application, NSDictionary* userInfo,void (^UIBackgroundFetchResult)(void) ) {
     NSLog(@"swizzled didReceiveRemoteNotification");
-
-   [[AppsFlyerTracker sharedTracker] handlePushNotification:userInfo];
-
+    
+    [[AppsFlyerLib shared] handlePushNotification:userInfo];
+    
     if(__original_didReceiveRemoteNotification_Imp){
         return ((BOOL(*)(id, SEL, UIApplication*, NSDictionary*, (UIBackgroundFetchResult)))__original_didReceiveRemoteNotification_Imp)(self, _cmd, application, userInfo, nil);
     }
     return YES;
 }
- */
 
-/** <remove comment if you are swizzling openURL>
+
+
 BOOL __swizzled_openURL(id self, SEL _cmd, UIApplication* application, NSURL* url, NSDictionary * options) {
     NSLog(@"swizzled openURL");
-    [[AppsFlyerTracker sharedTracker] handleOpenUrl:url options:options];
+    [[AppsFlyerLib shared] handleOpenUrl:url options:options];
     if(__original_openUrl_Imp){
         return ((BOOL(*)(id, SEL, UIApplication*, NSURL*, NSDictionary*))__original_openUrl_Imp)(self, _cmd, application, url, options);
     }
-    return YES;
+    return NO;
 }
-*/
+
 
 @end
-