+++ Assets/AppsFlyer/Plugins/iOS/AppsFlyeriOSWrapper.h
@@ -7,21 +7,22 @@
 
 #import "AFUnityUtils.mm"
 #import "UnityAppController.h"
-#if __has_include(<AppsFlyerLib/AppsFlyerTracker.h>)
-#import <AppsFlyerLib/AppsFlyerTracker.h>
+#if __has_include(<AppsFlyerLib/AppsFlyerLib.h>)
+#import <AppsFlyerLib/AppsFlyerLib.h>
 #else
-#import "AppsFlyerTracker.h"
+#import "AppsFlyerLib.h"
 #endif
 
-@interface AppsFlyeriOSWarpper : NSObject <AppsFlyerTrackerDelegate>
-
+@interface AppsFlyeriOSWarpper : NSObject <AppsFlyerLibDelegate, AppsFlyerDeepLinkDelegate>
++ (BOOL) didCallStart;
++ (void) setDidCallStart:(BOOL)val;
 @end
 
 
 static AppsFlyeriOSWarpper *_AppsFlyerdelegate;
 static const int kPushNotificationSize = 32;
 
-NSString* ConversionDataCallbackObject;
+static NSString* ConversionDataCallbackObject = @"AppsFlyerObject";
 
 static const char* VALIDATE_CALLBACK = "didFinishValidateReceipt";
 static const char* VALIDATE_ERROR_CALLBACK = "didFinishValidateReceiptWithError";
@@ -31,5 +32,15 @@ static const char* OAOA_CALLBACK = "onAppOpenAttribution";
 static const char* OAOA_ERROR_CALLBACK = "onAppOpenAttributionFailure";
 static const char* GENERATE_LINK_CALLBACK = "onInviteLinkGenerated";
 static const char* OPEN_STORE_LINK_CALLBACK = "onOpenStoreLinkGenerated";
+static const char* START_REQUEST_CALLBACK = "requestResponseReceived";
+static const char* IN_APP_RESPONSE_CALLBACK = "inAppResponseReceived";
+static const char* ON_DEEPLINKING = "onDeepLinking";
+
 
 static NSString* validateObjectName = @"";
+static NSString* openStoreObjectName = @"";
+static NSString* generateInviteObjectName = @"";
+static NSString* startRequestObjectName = @"";
+static NSString* inAppRequestObjectName = @"";
+static NSString* onDeeplinkingObjectName = @"";
+